requirements.txt
langchain
PyPDF2
python-dotenv
#streamlit==1.18.1
openai
faiss-cpu
altair
tiktoken
streamlit
htmlTemplate
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2